requirements.txt
fastapi
uvicorn
pandas
numpy
supabase
scikit-learn
python-dotenv
